requirements.txt
django
djangorestframework 
django-cors-headers 
pygments 
djangorestframework-simplejwt
psycopg2